+++ genweb/smartportlet/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
 from zope.i18nmessageid import MessageFactory
 
-DOMAIN = u'genweb.extras'
+DOMAIN = u'genweb.smartportlet'
 _ = MessageFactory(DOMAIN)