+++ genweb/smartportlet/locales/update.sh
@@ -1,4 +1,4 @@
-domain=genweb.extras
+domain=genweb.smartportlet
 i18ndude rebuild-pot --pot $domain.pot --create $domain --merge $domain-manual.pot ../
 i18ndude sync --pot $domain.pot */LC_MESSAGES/$domain.po
 domain=plone