+++ HISTORY.rst
@@ -2,5 +2,5 @@
 0.1 (unreleased)
 ----------------
 
-* Initial creation of genweb.extras
+* Initial creation of genweb.smartportlet
   Carles Bruguera